move editableTabTitle to groupsearchviewtabs folder
