+++ static/app/views/issueList/groupSearchViewTabs/editableTabTitle.tsx
